requirements.txt
pandas==3.0.1
customtkinter==5.2.2
requests==2.33.1
Pillow==12.2.0
